+++ stdnet/orm/__init__.py
@@ -3,3 +3,4 @@ from .models import *
 from .mapper import *
 from .fields import *
 from .std import *
+from .signals import *